feat: separate SetFullMode component
