+++ Frontend/lexica-next-front/src/components/sets/modes/SetFullMode.tsx
@@ -1,0 +1,361 @@
+import { useEffect, useState } from 'react';
+import { IconCheck, IconX } from '@tabler/icons-react';
+import { useNavigate } from 'react-router';
+import { Alert, Button, Container, Group, Paper, Progress, Radio, Stack, Text, TextInput, Title } from '@mantine/core';
+import { type EntryDto, type GetSetResponse } from '../../../hooks/api';
+
+interface FullModeEntry extends EntryDto {
+  englishCloseCounter: number;
+  nativeCloseCounter: number;
+  englishOpenCounter: number;
+  nativeOpenCounter: number;
+}
+
+type QuestionType = 'english-close' | 'native-close' | 'english-open' | 'native-open';
+
+interface Question {
+  entry: FullModeEntry;
+  entryIndex: number;
+  type: QuestionType;
+  question: string;
+  options?: string[];
+  correctAnswer: string;
+}
+
+export interface SetFullModeProps {
+  set: GetSetResponse;
+}
+
+export function SetFullMode({ set }: SetFullModeProps) {
+  const navigate = useNavigate();
+  const [entries, setEntries] = useState<FullModeEntry[]>([]);
+  const [currentQuestion, setCurrentQuestion] = useState<Question | null>(null);
+  const [userAnswer, setUserAnswer] = useState('');
+  const [showFeedback, setShowFeedback] = useState(false);
+  const [isCorrect, setIsCorrect] = useState(false);
+  const [isComplete, setIsComplete] = useState(false);
+
+  useEffect(() => {
+    if (set?.entries) {
+      const initialEntries = set.entries.map((entry) => ({
+        ...entry,
+        englishCloseCounter: 0,
+        nativeCloseCounter: 0,
+        englishOpenCounter: 0,
+        nativeOpenCounter: 0,
+      }));
+      setEntries(initialEntries);
+      generateNextQuestion(initialEntries);
+    }
+  }, [set]);
+
+  const generateNextQuestion = (currentEntries: FullModeEntry[]) => {
+    const shuffledEntries = [...currentEntries].sort(() => Math.random() - 0.5);
+    const selectedEntries = shuffledEntries.slice(0, Math.min(7, shuffledEntries.length));
+
+    const eligibleEntries = selectedEntries.filter((entry) => {
+      return (
+        entry.englishCloseCounter < 1 ||
+        entry.nativeCloseCounter < 1 ||
+        entry.englishOpenCounter < 2 ||
+        entry.nativeOpenCounter < 2
+      );
+    });
+
+    if (eligibleEntries.length === 0) {
+      const allComplete = currentEntries.every(
+        (entry) =>
+          entry.englishCloseCounter >= 1 &&
+          entry.nativeCloseCounter >= 1 &&
+          entry.englishOpenCounter >= 2 &&
+          entry.nativeOpenCounter >= 2,
+      );
+
+      if (allComplete) {
+        setIsComplete(true);
+        return;
+      }
+
+      generateNextQuestion(currentEntries);
+      return;
+    }
+
+    const selectedEntry = eligibleEntries[Math.floor(Math.random() * eligibleEntries.length)];
+    const entryIndex = currentEntries.findIndex((e) => e.word === selectedEntry.word);
+
+    const availableTypes: QuestionType[] = [];
+
+    if (selectedEntry.englishCloseCounter < 1) {
+      availableTypes.push('english-close');
+    }
+
+    if (selectedEntry.nativeCloseCounter < 1) {
+      availableTypes.push('native-close');
+    }
+
+    if (availableTypes.length === 0) {
+      if (selectedEntry.englishOpenCounter < 2) {
+        availableTypes.push('english-open');
+      }
+
+      if (selectedEntry.nativeOpenCounter < 2) {
+        availableTypes.push('native-open');
+      }
+    }
+
+    const questionType: QuestionType = availableTypes[Math.floor(Math.random() * availableTypes.length)];
+    const question = generateQuestion(selectedEntry, entryIndex, questionType, currentEntries);
+    setCurrentQuestion(question);
+  };
+
+  const generateQuestion = (
+    entry: FullModeEntry,
+    entryIndex: number,
+    type: QuestionType,
+    allEntries: FullModeEntry[],
+  ): Question => {
+    switch (type) {
+      case 'english-close': {
+        const correctTranslation = entry.translations?.[0] || '';
+        const wrongOptions = allEntries
+          .filter((e) => e.word !== entry.word)
+          .flatMap((e) => e.translations || [])
+          .filter((t) => t && t !== correctTranslation)
+          .slice(0, 3);
+
+        const options = [correctTranslation, ...wrongOptions].sort(() => Math.random() - 0.5);
+
+        return {
+          entry,
+          entryIndex,
+          type,
+          question: `What does "${entry.word}" mean?`,
+          options,
+          correctAnswer: correctTranslation,
+        };
+      }
+
+      case 'native-close': {
+        const correctWord = entry.word;
+        const wrongWords = allEntries
+          .filter((e) => e.word !== entry.word)
+          .map((e) => e.word || '')
+          .filter((word) => word !== '')
+          .slice(0, 3);
+
+        const wordOptions = [correctWord || '', ...wrongWords].sort(() => Math.random() - 0.5);
+
+        return {
+          entry,
+          entryIndex,
+          type,
+          question: `What is the English word for "${entry.translations?.[0] || ''}"?`,
+          options: wordOptions,
+          correctAnswer: correctWord || '',
+        };
+      }
+
+      case 'english-open':
+        return {
+          entry,
+          entryIndex,
+          type,
+          question: `What does "${entry.word}" mean? (Type your answer)`,
+          correctAnswer: entry.translations?.[0] || '',
+        };
+
+      case 'native-open':
+        return {
+          entry,
+          entryIndex,
+          type,
+          question: `What is the English word for "${entry.translations?.[0] || ''}"? (Type your answer)`,
+          correctAnswer: entry.word || '',
+        };
+
+      default:
+        throw new Error('Invalid question type');
+    }
+  };
+
+  const checkAnswer = () => {
+    if (!currentQuestion) {
+      return;
+    }
+
+    const correct = userAnswer.trim().toLowerCase() === currentQuestion.correctAnswer.toLowerCase();
+    setIsCorrect(correct);
+    setShowFeedback(true);
+
+    const updatedEntries = [...entries];
+    const entry = updatedEntries[currentQuestion.entryIndex];
+
+    if (correct) {
+      switch (currentQuestion.type) {
+        case 'english-close':
+          entry.englishCloseCounter += 1;
+          break;
+        case 'native-close':
+          entry.nativeCloseCounter += 1;
+          break;
+        case 'english-open':
+          entry.englishOpenCounter += 1;
+          break;
+        case 'native-open':
+          entry.nativeOpenCounter += 1;
+          break;
+      }
+    } else {
+      entry.englishCloseCounter = 0;
+      entry.nativeCloseCounter = 0;
+      entry.englishOpenCounter = 0;
+      entry.nativeOpenCounter = 0;
+    }
+
+    setEntries(updatedEntries);
+  };
+
+  const nextQuestion = () => {
+    setShowFeedback(false);
+    setUserAnswer('');
+    generateNextQuestion(entries);
+  };
+
+  const getProgress = () => {
+    const totalRequired = entries.length * 6;
+    const currentProgress = entries.reduce((sum, entry) => {
+      return (
+        sum +
+        Math.min(entry.englishCloseCounter, 1) +
+        Math.min(entry.nativeCloseCounter, 1) +
+        Math.min(entry.englishOpenCounter, 2) +
+        Math.min(entry.nativeOpenCounter, 2)
+      );
+    }, 0);
+
+    return totalRequired > 0 ? (currentProgress / totalRequired) * 100 : 0;
+  };
+
+  const handleKeyDown = (event: React.KeyboardEvent) => {
+    if (event.key === 'Enter' && userAnswer.trim() && !showFeedback) {
+      checkAnswer();
+    }
+  };
+
+  if (isComplete) {
+    return (
+      <>
+        <Container size="md">
+          <Stack gap="lg" align="center" py="xl">
+            <Title order={1} ta="center" c="green" fz={{ base: 'h2', md: 'h1' }}>
+              🎉 Congratulations!
+            </Title>
+            <Text fz={{ base: 'md', md: 'lg' }} ta="center">
+              You've completed the full mode for "{set?.name}"!
+            </Text>
+            <Text c="dimmed" ta="center" fz={{ base: 'sm', md: 'md' }}>
+              You've mastered all the words in this set through comprehensive practice.
+            </Text>
+            <Group wrap="wrap" justify="center">
+              <Button variant="light" onClick={() => navigate('/sets')} size="md">
+                Back to Sets
+              </Button>
+              <Button onClick={() => window.location.reload()} size="md">
+                Practice Again
+              </Button>
+            </Group>
+          </Stack>
+        </Container>
+      </>
+    );
+  }
+
+  if (!currentQuestion) {
+    return (
+      <>
+        <Container size="md">
+          <Alert color="orange" title="No questions available">
+            Unable to generate questions for this set.
+          </Alert>
+        </Container>
+      </>
+    );
+  }
+
+  return (
+    <>
+      <Stack gap="lg">
+        <Progress value={getProgress()} size="lg" radius="md" />
+
+        <Paper>
+          <Stack gap="lg">
+            <div>
+              <Text fz={{ base: 'md', md: 'lg' }} fw={600} mb="md">
+                {currentQuestion.question}
+              </Text>
+              <Text size="sm" c="dimmed">
+                Word type: {currentQuestion.entry.wordType}
+              </Text>
+            </div>
+
+            {!showFeedback ? (
+              <Stack gap="md">
+                {currentQuestion.options ? (
+                  <Radio.Group value={userAnswer} onChange={setUserAnswer}>
+                    <Stack gap="sm">
+                      {currentQuestion.options.map((option, index) => (
+                        <Radio key={index} value={option} label={option} size="md" />
+                      ))}
+                    </Stack>
+                  </Radio.Group>
+                ) : (
+                  <TextInput
+                    placeholder="Type your answer..."
+                    value={userAnswer}
+                    onChange={(e) => setUserAnswer(e.target.value)}
+                    size="lg"
+                    onKeyDown={handleKeyDown}
+                    autoFocus
+                  />
+                )}
+
+                <Button size="lg" onClick={checkAnswer} disabled={!userAnswer.trim()}>
+                  Check Answer
+                </Button>
+              </Stack>
+            ) : (
+              <Stack gap="md">
+                <Alert
+                  color={isCorrect ? 'green' : 'red'}
+                  icon={isCorrect ? <IconCheck size={16} /> : <IconX size={16} />}
+                  title={isCorrect ? 'Correct!' : 'Incorrect'}>
+                  {!isCorrect && (
+                    <Text>
+                      The correct answer is: <strong>{currentQuestion.correctAnswer}</strong>
+                    </Text>
+                  )}
+                </Alert>
+
+                <div>
+                  <Text fw={600} fz={{ base: 'md', md: 'lg' }}>
+                    {currentQuestion.entry.word}
+                  </Text>
+                  <Text c="dimmed" size="sm">
+                    ({currentQuestion.entry.wordType})
+                  </Text>
+                  <Text mt="sm" fz={{ base: 'sm', md: 'md' }}>
+                    <strong>Translations:</strong> {(currentQuestion.entry.translations || []).join(', ')}
+                  </Text>
+                </div>
+
+                <Button size="lg" onClick={nextQuestion} autoFocus>
+                  Continue
+                </Button>
+              </Stack>
+            )}
+          </Stack>
+        </Paper>
+      </Stack>
+    </>
+  );
+}
